requirements.txt
# Core Frameworks
crewai==0.28.8
langchain
langchain-community
langchain-openai
langchain-groq

# Vector Database & Embeddings
chromadb
langchain-chroma
langchain-huggingface
sentence-transformers
torch

# Data Processing & Extraction
pypdf
pymupdf
langchain-experimental
pandas
numpy

# Utilities
python-dotenv
tqdm